split qlog trace writer and QUIC qlog events into separate packages
