+++ qlog/interface.go
@@ -1,0 +1,14 @@
+package qlog
+
+import (
+	"io"
+)
+
+type Trace interface {
+	AddProducer() Recorder
+}
+
+type Recorder interface {
+	RecordEvent(Event)
+	io.Closer
+}
