+++ qlog/writer.go
@@ -89,7 +89,7 @@ func (t *FileSeq) AddProducer() *Writer {
 	}
 }
 
-func (t *FileSeq) record(eventTime time.Time, details eventDetails) error {
+func (t *FileSeq) record(eventTime time.Time, details Event) error {
 	t.mx.Lock()
 
 	if t.closed {
@@ -100,7 +100,7 @@ func (t *FileSeq) record(eventTime time.Time, details eventDetails) error {
 
 	t.events <- event{
 		RelativeTime: eventTime.Sub(t.referenceTime),
-		eventDetails: details,
+		Event:        details,
 	}
 	return nil
 }
@@ -158,7 +158,7 @@ func (w *Writer) Close() error {
 	return nil
 }
 
-func (w *Writer) RecordEvent(time time.Time, details eventDetails) {
-	err := w.t.record(time, details)
+func (w *Writer) RecordEvent(ev Event) {
+	err := w.t.record(time.Now(), ev)
 	_ = err
 }
