+++ qlogevents/frame.go
@@ -1,4 +1,4 @@
-package qlog
+package qlogevents
 
 import (
 	"fmt"
