+++ qlogevents/packet_header_test.go
@@ -1,4 +1,4 @@
-package qlog
+package qlogevents
 
 import (
 	"bytes"
